Delete place validation

+++ src/server/controllers/placesControllers/placesControllers.js
@@ -81,6 +81,16 @@ const deletePlace = async (req, res, next) => {
   const { placeId } = req.params;
 
   try {
+    const place = await Place.findById(placeId);
+
+    if (!place) {
+      return next(customError(404, "Bad request", "Place id not found"));
+    }
+
+    if (place.creator !== userId) {
+      return next(customError(403, "You can't perform this operation."));
+    }
+
     const deletedPlace = await Place.findByIdAndDelete(placeId);
 
     if (deletedPlace) {
